modules/voice-modulation/voice_processor.py
import os
import numpy as np
import soundfile as sf
import librosa
import tensorflow as tf
from tensorflow import keras
from typing import Dict, List, Optional, Tuple
import logging
import uuid
import webrtc_noise_suppression as ns
import sounddevice as sd

class VoiceModulator:
    """
    Advanced Voice Modulation System for Minecraft Voice Communication
    
    Features:
    - Real-time voice processing
    - Noise reduction
    - Voice style transfer
    - Emotion detection
    - Privacy-preserving voice anonymization
    """
    
    def __init__(
        self, 
        models_dir: str = '/opt/mlsp/voice_models', 
        cache_dir: str = '/tmp/mlsp_voice_cache'
    ):
        """
        Initialize Voice Modulation System
        
        :param models_dir: Directory for storing ML models
        :param cache_dir: Temporary directory for audio processing
        """
        self.models_dir = models_dir
        self.cache_dir = cache_dir
        
        # Create necessary directories
        os.makedirs(models_dir, exist_ok=True)
        os.makedirs(cache_dir, exist_ok=True)
        
        # Configure logging
        self.logger = logging.getLogger('MLSPVoiceModulator')
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        )
        
        # Load pre-trained models
        self._load_models()
    
    def _load_models(self):
        """
        Load pre-trained voice processing models
        
        Models:
        1. Noise Reduction Model
        2. Voice Style Transfer Model
        3. Emotion Detection Model
        4. Voice Anonymization Model
        """
        try:
            # Noise Reduction Model (Placeholder)
            noise_reduction_path = os.path.join(self.models_dir, 'noise_reduction.h5')
            if os.path.exists(noise_reduction_path):
                self.noise_reduction_model = keras.models.load_model(noise_reduction_path)
            else:
                self.noise_reduction_model = self._create_noise_reduction_model()
            
            # Voice Style Transfer Model (Placeholder)
            style_transfer_path = os.path.join(self.models_dir, 'style_transfer.h5')
            if os.path.exists(style_transfer_path):
                self.style_transfer_model = keras.models.load_model(style_transfer_path)
            else:
                self.style_transfer_model = self._create_style_transfer_model()
            
            # Emotion Detection Model (Placeholder)
            emotion_model_path = os.path.join(self.models_dir, 'emotion_detection.h5')
            if os.path.exists(emotion_model_path):
                self.emotion_model = keras.models.load_model(emotion_model_path)
            else:
                self.emotion_model = self._create_emotion_detection_model()
            
            # Voice Anonymization Model (Placeholder)
            anonymization_path = os.path.join(self.models_dir, 'voice_anonymization.h5')
            if os.path.exists(anonymization_path):
                self.anonymization_model = keras.models.load_model(anonymization_path)
            else:
                self.anonymization_model = self._create_anonymization_model()
        
        except Exception as e:
            self.logger.error(f"Model loading error: {e}")
            raise
    
    def _create_noise_reduction_model(self) -> keras.Model:
        """
        Create a basic neural network for noise reduction
        
        :return: Keras noise reduction model
        """
        model = keras.Sequential([
            keras.layers.Input(shape=(None, 1)),
            keras.layers.Conv1D(32, 3, activation='relu', padding='same'),
            keras.layers.MaxPooling1D(2),
            keras.layers.Conv1D(64, 3, activation='relu', padding='same'),
            keras.layers.MaxPooling1D(2),
            keras.layers.GlobalAveragePooling1D(),
            keras.layers.Dense(1, activation='sigmoid')
        ])
        
        model.compile(
            optimizer='adam',
            loss='binary_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def _create_style_transfer_model(self) -> keras.Model:
        """
        Create a voice style transfer model
        
        :return: Keras style transfer model
        """
        model = keras.Sequential([
            keras.layers.Input(shape=(None, 1)),
            keras.layers.LSTM(128, return_sequences=True),
            keras.layers.LSTM(64, return_sequences=True),
            keras.layers.Dense(1, activation='tanh')
        ])
        
        model.compile(
            optimizer='adam',
            loss='mean_squared_error'
        )
        
        return model
    
    def _create_emotion_detection_model(self) -> keras.Model:
        """
        Create an emotion detection model
        
        :return: Keras emotion detection model
        """
        model = keras.Sequential([
            keras.layers.Input(shape=(None, 1)),
            keras.layers.Conv1D(64, 3, activation='relu', padding='same'),
            keras.layers.MaxPooling1D(2),
            keras.layers.Conv1D(128, 3, activation='relu', padding='same'),
            keras.layers.GlobalAveragePooling1D(),
            keras.layers.Dense(64, activation='relu'),
            keras.layers.Dense(5, activation='softmax')  # 5 emotion classes
        ])
        
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def _create_anonymization_model(self) -> keras.Model:
        """
        Create a voice anonymization model
        
        :return: Keras voice anonymization model
        """
        model = keras.Sequential([
            keras.layers.Input(shape=(None, 1)),
            keras.layers.Conv1D(64, 3, activation='relu', padding='same'),
            keras.layers.MaxPooling1D(2),
            keras.layers.Conv1D(128, 3, activation='relu', padding='same'),
            keras.layers.GlobalAveragePooling1D(),
            keras.layers.Dense(64, activation='relu'),
            keras.layers.Dense(1, activation='tanh')
        ])
        
        model.compile(
            optimizer='adam',
            loss='mean_squared_error'
        )
        
        return model
    
    def reduce_noise(self, audio_data: np.ndarray) -> np.ndarray:
        """
        Reduce background noise from audio
        
        :param audio_data: Input audio numpy array
        :return: Noise-reduced audio
        """
        try:
            # WebRTC noise suppression
            denoised_audio = ns.reduce_noise(audio_data)
            
            # Optional ML-based refinement
            if self.noise_reduction_model:
                # Reshape and normalize
                audio_normalized = audio_data / np.max(np.abs(audio_data))
                audio_input = audio_normalized.reshape(1, -1, 1)
                
                # Predict noise mask
                noise_mask = self.noise_reduction_model.predict(audio_input)
                denoised_audio = audio_data * (1 - noise_mask)
            
            return denoised_audio
        
        except Exception as e:
            self.logger.error(f"Noise reduction error: {e}")
            return audio_data
    
    def transfer_voice_style(
        self, 
        source_audio: np.ndarray, 
        target_style: str = 'robotic'
    ) -> np.ndarray:
        """
        Apply voice style transfer
        
        :param source_audio: Input audio numpy array
        :param target_style: Desired voice style
        :return: Style-transferred audio
        """
        try:
            # Normalize audio
            audio_normalized = source_audio / np.max(np.abs(source_audio))
            audio_input = audio_normalized.reshape(1, -1, 1)
            
            # Style transfer
            if self.style_transfer_model:
                style_vector = {
                    'robotic': np.array([1.0]),
                    'deep': np.array([-1.0]),
                    'high_pitch': np.array([0.5]),
                    'low_pitch': np.array([-0.5])
                }.get(target_style, np.array([0.0]))
                
                transferred_audio = self.style_transfer_model.predict(audio_input) * style_vector
                return transferred_audio.flatten()
            
            return source_audio
        
        except Exception as e:
            self.logger.error(f"Voice style transfer error: {e}")
            return source_audio
    
    def detect_emotion(self, audio_data: np.ndarray) -> Dict[str, float]:
        """
        Detect emotion from audio
        
        :param audio_data: Input audio numpy array
        :return: Emotion probabilities
        """
        try:
            # Extract audio features
            mfccs = librosa.feature.mfcc(y=audio_data, n_mfcc=13)
            
            # Prepare input
            audio_input = mfccs.T.reshape(1, -1, 13)
            
            # Predict emotions
            if self.emotion_model:
                emotion_probs = self.emotion_model.predict(audio_input)[0]
                emotions = ['neutral', 'happy', 'sad', 'angry', 'surprised']
                
                return dict(zip(emotions, emotion_probs))
            
            return {}
        
        except Exception as e:
            self.logger.error(f"Emotion detection error: {e}")
            return {}
    
    def anonymize_voice(self, audio_data: np.ndarray) -> np.ndarray:
        """
        Anonymize voice to protect identity
        
        :param audio_data: Input audio numpy array
        :return: Anonymized audio
        """
        try:
            # Normalize audio
            audio_normalized = audio_data / np.max(np.abs(audio_data))
            audio_input = audio_normalized.reshape(1, -1, 1)
            
            # Anonymization
            if self.anonymization_model:
                anonymization_mask = self.anonymization_model.predict(audio_input)
                anonymized_audio = audio_data * anonymization_mask
                return anonymized_audio.flatten()
            
            return audio_data
        
        except Exception as e:
            self.logger.error(f"Voice anonymization error: {e}")
            return audio_data
    
    def process_audio(
        self, 
        audio_data: np.ndarray, 
        config: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Comprehensive audio processing
        
        :param audio_data: Input audio numpy array
        :param config: Processing configuration
        :return: Processed audio and metadata
        """
        config = config or {}
        
        # Process audio
        denoised_audio = self.reduce_noise(audio_data)
        
        # Optional style transfer
        if config.get('style_transfer'):
            denoised_audio = self.transfer_voice_style(
                denoised_audio, 
                config.get('style', 'robotic')
            )
        
        # Emotion detection
        emotions = self.detect_emotion(denoised_audio)
        
        # Anonymization
        if config.get('anonymize', False):
            denoised_audio = self.anonymize_voice(denoised_audio)
        
        # Generate unique audio ID
        audio_id = str(uuid.uuid4())
        
        # Save processed audio
        output_path = os.path.join(
            self.cache_dir, 
            f'processed_audio_{audio_id}.wav'
        )
        sf.write(output_path, denoised_audio, 44100)
        
        return {
            'audio_id': audio_id,
            'processed_audio_path': output_path,
            'emotions': emotions,
            'original_duration': len(audio_data) / 44100,  # seconds
            'processed_duration': len(denoised_audio) / 44100  # seconds
        }

def main():
    """Example usage of Voice Modulator"""
    voice_modulator = VoiceModulator()
    
    try:
        # Simulate audio recording
        duration = 5  # seconds
        sample_rate = 44100
        audio_data = sd.rec(
            int(duration * sample_rate), 
            samplerate=sample_rate, 
            channels=1, 
            dtype='float64'
        )
        sd.wait()
        
        # Process audio
        result = voice_modulator.process_audio(
            audio_data.flatten(), 
            config={
                'style_transfer': True,
                'style': 'robotic',
                'anonymize': True
            }
        )
        
        print("Voice Processing Result:")
        print(f"Audio ID: {result['audio_id']}")
        print(f"Processed Audio Path: {result['processed_audio_path']}")
        print("Detected Emotions:", result['emotions'])
    
    except Exception as e:
        print(f"Voice modulation error: {e}")

if __name__ == '__main__':
    main()
